+++ release-notes.md
@@ -1,3 +1,16 @@
+# Mercury 13.2.0 #
+
+## Add create recorded entry action ##
+
+- Issue Type: Feature
+- Issue ID: PLAT-7827
+
+### configuration ###
+None
+
+### Deployment scripts ###
+	php deployment/updates/scripts/add_permissions/2017_08_06_live_stream_add_create_recording_across_dc.php
+
 # Mercury 13.1.0 #
 
 ## Add new Search plugin ##
@@ -11,6 +24,7 @@ Add the following to plugins.ini file: "Search"
 ### Deployment scripts ###
     1. php /opt/kaltura/app/deployment/base/scripts/installPlugins.php
     2. mysql -h{HOSTNAME} -u{USER} -p{PASSWORD} kaltura_sphinx_log < /opt/kaltura/app/deployment/updates/sql/2017_05_15_add_type_column_sphinx_log.sql
+
 
 ## Preserve Aspect Ratio accurately  ##